req18 file view function
the view file button on sidebar  now redirects to a separate file view page. files show preview image and filename, expand when clicked, and are sorted and separated by project. currently, contractor can view files regardless of who uploaded them. oddly, files are not showing for subcontractors even if they are the user that uploaded them. that is up next.