Subcontractor Show Completed Tasks Fixed!
DOM rendering caused the issue on the subcontractor side... How and when it was rendered. The tab system makes on tab visible at a time and by default, information tab is active when page loads! This made subcontractor tab hidden, like the #toggle-completed. DOMContentLoaded will therefore not show the button b/c it can not find it! Delaying the toggle functionality until this tab is active helped solve the issues. I thought it was the conditional statements, but after a lot of digging and ChatGPT prompting, this was the issue. Goes to show the incredible usage of AI in debugging.